Add checks for file exists.

+++ scripts/pains/LillyMedchemRules/pains.py
@@ -11,8 +11,10 @@ def run_pains_filter(inputs):
   filtered_files = ['bad0.smi', 'bad1.smi', 'bad2.smi', 'bad3.smi']
   os.chdir(os.path.abspath(os.path.dirname(__file__)))
 
-  os.remove("okmedchem.smi")
-  os.remove("input.smi")
+  if os.path.exists("okmedchem.smi"):
+    os.remove("okmedchem.smi")
+  if os.path.exists("input.smi"):
+    os.remove("input.smi")
 
   for filtered_file in filtered_files:
     os.remove(filtered_file)
